requirements.txt
Django==4.1
django-cors-headers==3.13.0
django-extensions==3.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
Pillow==9.2.0
PyMySQL==1.0.2
pytz==2022.2.1
requests==2.28.1